+++ tests/integration_tests.rs
@@ -789,3 +789,20 @@ fn test_filter_apply() {
     .join()
     .unwrap();
 }
+
+#[test]
+fn test_u16_overflow() {
+    let rule = sock_filter {
+        code: 6,
+        jt: 0,
+        jf: 0,
+        k: 0,
+    };
+    let rules = vec![rule; 0x10_000];
+    let err = seccompiler::apply_filter(&rules).unwrap_err();
+    assert!(matches!(err, seccompiler::Error::TooManyInstructions));
+    assert_eq!(
+        err.to_string(),
+        "Cannot install filter with more than 65535 instructions."
+    );
+}